==> 1. External Data Collection.ipynb <==
# <font color='CC3D3D'> 0.1 External Data Collection
    
- 일별 KOSPI 지수 크롤링
- **KO**rea Composite **S**tock **P**rice **I**ndex의 약어로 우리나라의 주가 동향을 대표하는 Index  
- 증권 거래소에 상장된 종목들의 주식 가격을 종합적으로 표시한 수치로 시장 전체의 주가 움직임을 이해할 수 있는 지표
- 1981년부터 2022년 데이터까지 수집한 데이터라 2022년 2월 28일 이전의 데이터는 CSV에서 제거
	- <span style="color:blue"> **Kospi_Index.csv** </span> 생성

## Import

import FinanceDataReader as fdr
import pandas as pd

fdr.__version__

Close : 종가  
Open : 시가  
High : 고가  
Low : 저가  
Volume : 거래량  
Change : 등락률(전일대비 등락률)  
➡️ 등락률 = (현재 종가 - 전일 종가) / 전일종가  

## Data Load

# 존재하는 첫 시점부터 마지막으로 장이 끝난 시점까지의 지수 데이터 다운로드

kospi = fdr.DataReader('KS11')

kospi.info()

kospi.head()

kospi.tail()

# Deployment CSV

# 2022-02-22 일부터 선정
kospi.to_csv('../Data/Kospi_Index.csv')

- 수집한 데이터가 1981년부터 2022년 데이터까지 수집한 데이터를 CSV에서 제거
