==> 2. Extract Applied Or Not.ipynb <==
# <font color='CC3D3D'> 0.2 Extract_Applied_Or_Not
    
- loan_result.csv와 user_spec.csv 파일을 합쳐 application id에 loan id를 달아줌
- 시간 정보를 추출하여 유저별로 각 일자에 대출 신청을 했는지 하지 않았는지 구분
    - <span style="color:blue"> **loan_apply.csv** </span> 생성

# Import

import os
import gc
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib as mpl  # 기본 설정 만지는 용도
import matplotlib.pyplot as plt
import matplotlib.font_manager as fm  # 폰트 관련 용도
from IPython.display import display
import warnings

## for dimension reduction or feature selection
from sklearn.decomposition import PCA
from sklearn.linear_model import Lasso

warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
%matplotlib inline

# Data Load

loan = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/loan_result.csv", engine="python")
user = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/user_spec.csv", engine="python")

user[user['user_id'] == 879698]

# Log Data Preprocessing

def AddDate(df:pd.DataFrame, key:str, add_day:bool=False, add_time:bool=False) -> None:
    df[key] = pd.to_datetime(df[key])
    df[f"{key}_year"] = df[key].dt.year
    df[f"{key}_month"] = df[key].dt.month
    if add_day:
        df[f"{key}_day"] = df[key].dt.day
    if add_time:
        df[f"{key}_hour"] = df[key].dt.hour
        df[f"{key}_min"] = df[key].dt.minute
        df[f"{key}_sec"] = df[key].dt.second

* apply한 날을 일단 추출 - 6월 이전의 데이터에 대해서만

loan['insert_date'] = pd.to_datetime(loan['loanapply_insert_time']).dt.date

loan.info()

loan['month'] = pd.to_datetime(loan['insert_date']).dt.month
loan = loan[loan['month'] < 6]
loan.drop('month', axis=1, inplace=True)

loan.isnull().sum()

app_user = {user['application_id'].iloc[i]: user['user_id'].iloc[i] for i in range(user.shape[0])}
len(app_user)

user['application_id'].nunique()

loan['application_id'].nunique()

loan['user_id'] = loan['application_id'].map(app_user)
loan.head()

loan.isnull().sum()

* 113명의 유저 아이디가 loan의 application id와 매핑되지 않는다!

user.isnull().sum()

app_nulls = loan[loan['user_id'].isna()]
app_nulls['insert_date'] = pd.to_datetime(app_nulls['insert_date'])

app_nulls.sort_values(by='insert_date', inplace=True)
app_nulls.reset_index(drop=True, inplace=True)

app_nulls.tail()

loan.dropna(subset=['user_id'], inplace=True)
loan.isnull().sum()

loan_apply = loan[['user_id', 'insert_date', 'is_applied']]
display(loan_apply)

loan_apply.drop_duplicates(inplace=True)
display(loan_apply)

loan_apply.sort_values(by=['user_id', 'insert_date'], inplace=True)
loan_apply.reset_index(drop=True, inplace=True)
loan_apply

loan_total = loan_apply.groupby(by=['user_id', 'insert_date']).sum()
loan_total

loan_total.reset_index(drop=False, inplace=True)
loan_total

loan_total['is_applied'] = loan_total['is_applied'].map(lambda x: x > 0).astype('int')
loan_total.head()

loan_total.info()

# Deployment CSV

loan_total.to_csv('../Data/loan_apply.csv', index=False)
