==> 3. Merge Loan Log.ipynb <==
# <font color='CC3D3D'> 0.3 Merge_Loan_Log
    
- log_data.csv와 loan_apply.csv 파일을 합쳐 로그 데이터에 개별 유저가 대출 신청을 했는지 여부 구분
- 대출 신청을 실제로 한 데이터
		- <span style="color:blue"> **log_applied.csv** </span> 생성
- 대출 신청을 실제로 하지 않은 데이터
    - <span style="color:blue"> **log_non_applied.csv** </span> 생성
- 위의 두 데이터를 합친 데이터
		- <span style="color:blue"> **log_applied_history.csv** </span> 생성

# Import

import os
import gc
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib as mpl  # 기본 설정 만지는 용도
import matplotlib.pyplot as plt
import matplotlib.font_manager as fm  # 폰트 관련 용도
from IPython.display import display
import warnings

## for dimension reduction or feature selection
from sklearn.decomposition import PCA
from sklearn.linear_model import Lasso

warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
%matplotlib inline

# Data Load

log = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/log_data.csv", engine="python")
apply_date = pd.read_csv("../Data/loan_apply.csv", engine="python")

def AddDate(df:pd.DataFrame, key:str, add_day:bool=False, add_time:bool=False) -> None:
    df[key] = pd.to_datetime(df[key])
    df[f"{key}_year"] = df[key].dt.year
    df[f"{key}_month"] = df[key].dt.month
    if add_day:
        df[f"{key}_day"] = df[key].dt.day
    if add_time:
        df[f"{key}_hour"] = df[key].dt.hour
        df[f"{key}_min"] = df[key].dt.minute
        df[f"{key}_sec"] = df[key].dt.second

* apply한 날을 일단 추출

log.info()

apply_date.info()

# Log Engineering

log.drop(['mp_os', 'mp_app_version'], axis=1, inplace=True)
log['date_cd'] = pd.to_datetime(log['date_cd']).dt.date
apply_date['insert_date'] = pd.to_datetime(apply_date['insert_date']).dt.date

display(log['user_id'].nunique())
display(apply_date['user_id'].nunique())

log.head()

apply_date.head()

log.sort_values(by=['user_id', 'date_cd'], inplace=True)
log.reset_index(drop=True, inplace=True)
log.head()

apply_date.sort_values(by=['user_id', 'insert_date'], inplace=True)
apply_date.reset_index(drop=True, inplace=True)
apply_date.columns = ['user_id', 'date_cd', 'is_applied']
apply_date.head()

left = pd.merge(left=log, right=apply_date, on=['user_id', 'date_cd'], how='left')
display(left.info())
left.drop_duplicates(inplace=True)
display(left)

left_applied = left[left['is_applied'] > 0]
right_applied = left[left['is_applied'] == 0]

left_applied.info()
right_applied.info()

# Deployment CSV

left_applied.to_csv("../Data/log_applied.csv", index=False)
right_applied.to_csv("../Data/log_non_applied.csv", index=False)
left.to_csv('../Data/log_applied_history.csv', index=False)
