==> 5.1. Clean Data &  External Data Merge.ipynb <==
# <font color='CC3D3D'> Clean Data & External Data Merge
    
### 1. Data Merge
- Clansing한 <span style="color:green">**user_spec_clansing.csv, loan_result_clansing, log_data_clansing.csv** </span>와 외부데이터 <span style="color:green">**소비자동향조사.csv, 소비자물가지수.csv', USA lending rate.csv, korea lending rate.csv, 성_연령별_실업률.csv, 성_연령별_실업자.csv, Kospi_Index.csv, crypto_currency_price_mod.csv** </span> Merge 진행

### 2. Deployment CSV
- Clansing한 원본 데이터와 외부데이터를 merge한 <span style="color:blue">**Master_df.csv** </span>을 생성

# Import

import pandas as pd
import numpy as np
import re

import matplotlib.pyplot as plt
import seaborn as sns

# Data Load

user_spec = pd.read_csv('../Data/user_spec_clansing.csv')
loan_result = pd.read_csv('../Data/loan_result_clansing.csv')
log_data = pd.read_csv('../Data/log_data_clansing.csv')

display(user_spec)

display(loan_result)

display(log_data)

user_spec.isna().sum()

loan_result.isna().sum()

log_data.isna().sum()

user_spec.info()

loan_result.info()

log_data.info()

# Data Merge

## loan_result & user_spec

master_df = pd.merge(loan_result,user_spec,on='application_id',how='left')

target = master_df.is_applied
master_df.drop(['is_applied'],axis='columns',inplace=True)
master_df['is_applied'] = target

master_df

master_df.isna().sum()

master_df = master_df.dropna(subset = ['user_id'])

master_df.isna().sum()

# 외부 데이터 Merge

csi_index = pd.read_csv('../Data/소비자동향조사.csv',encoding = 'CP949' ).copy()
price = pd.read_csv('../Data/소비자물가지수.csv',encoding = 'CP949').copy()
us_lending = pd.read_csv('../Data/USA lending rate.csv').copy()
k_lending = pd.read_csv('../Data/korea lending rate.csv').copy()
unemployment_rate = pd.read_csv('../Data/성_연령별_실업률.csv',encoding = 'CP949').copy()
unemployment_count = pd.read_csv('../Data/성_연령별_실업자.csv',encoding = 'CP949').copy()
kospi = pd.read_csv('../Data/Kospi_Index.csv').copy()
coin = pd.read_csv('../Data/crypto_currency_price_mod.csv').copy()

## csi_index

csi_index.head()

csi_index.dtypes

- 소비심리지수 만들어야 함
- 병합시 기준 고려

## price

price = price.T.rename(columns={0: 'price_index'}).drop('시도별')
price = price.reset_index().rename(columns={'index': 'Date'})
price.Date = price.Date.apply(lambda x : x[:7])

price.price_index =  price.price_index.astype('float64')

## us_lending

# 열이름 변경 
us_lending.rename(columns = {'lending rate': 'us_lending_rate'}, inplace = True)

## k_lending

# 열이름 변경 
k_lending.rename(columns = {'lending rate': 'k_lending_rate'}, inplace = True)

## unemployment_rate

unemployment_rate.Date = unemployment_rate.Date.apply(lambda x : x[:7])

## unemployment_count

unemployment_count.Date = unemployment_count.Date.apply(lambda x : str(x)[:7])

## kospi

- csv에서 하루씩 밀었음

kospi.head()

#### 코스피에도 비어있는 값이 있어 처리가 필요하겠다.

주말과 휴일에는 코스피가 동결되어 있겠지?

-> 따라서 전날 코스피로 채움   

data leakage에 걸리지 않게 하루씩 미루기로 했었나?

## coin

- csv에서 하루씩 밀었음

coin['btc_price'] = coin.btc_price.apply(lambda x: re.sub(r"[^0-9]", "", str(x)))
coin['eth_price'] = coin.eth_price.apply(lambda x: re.sub(r"[^0-9]", "", str(x)))

coin['btc_price'] = pd.to_numeric(coin['btc_price'])
coin['eth_price'] = pd.to_numeric(coin['eth_price'])

coin.date

# Loan_result와 날짜 기준으로 병합

# 병합을 위한 key 생성
master_df['month_day'] = master_df.loanapply_insert_time.apply(lambda x: x[:10])
master_df['month'] = master_df.loanapply_insert_time.apply(lambda x: x[:7])

master_df

## 외부데이터 병합 

# price_index 병합
master_df = pd.merge(master_df, price, left_on='month',right_on = 'Date', how = 'left').drop('Date', axis = 1)


# lending 정보 결합
master_df = pd.merge(master_df, k_lending, left_on='month_day', right_on ='Date',how = 'left').drop('Date', axis = 1)
master_df = pd.merge(master_df, us_lending, left_on='month_day', right_on ='Date',how = 'left').drop('Date', axis = 1)

# coin 정보 결합
master_df = pd.merge(master_df, coin, left_on='month_day', right_on ='date',how = 'left').drop('date', axis = 1)

# kospi 정보 결합
master_df = pd.merge(master_df, kospi, left_on='month_day', right_on ='Date',how = 'left').drop('Date', axis = 1)

master_df

tp_master = master_df

# 나이컬럼 만들기
tp_master['나이'] = 2022 - tp_master.birth_year + 1

# 함수 생성
def age_cat(x):
    if x < 19:
        return '15 - 19세'
    elif x >= 20 and x < 30:
        return '20 - 29세'
    elif x >= 30 and x < 40:
        return '30 - 39세'
    elif x >= 40 and x < 50:
        return '40 - 49세'
    elif x >= 50 and x < 60:
        return '50 - 59세'
    elif x >= 60:
        return '60세이상'

# 연령대로 변환하기
tp_master['연령계층별'] = tp_master['나이'].apply(age_cat)

## 날짜 / 성별 / 나이로 DF들 병합하기
# 실업자 수 병합
tp_master = pd.merge(tp_master, unemployment_count, left_on=['month', 'gender', '연령계층별'], right_on = ['Date', '성별', '연령계층별'], how = 'left').drop(['Date', '성별'], axis = 1)

# 실업률 병합
tp_master = pd.merge(tp_master, unemployment_rate, left_on=['month', 'gender', '연령계층별'], right_on = ['Date', '성별', '연령계층별'], how = 'left').drop(['Date', '성별'], axis = 1)

tp_master.info()

tp_master

tp_master를 만들면 연산량이 어마어마하네.... 따로 떼서 만들고 다시 붙여서 외부데이터끼리    
상관관계랑 기존 데이터 간의 상관관계도 봐야할것 같은데 그러면 데이터가 또 존나커져서 돌아갈까...


# csi지수 병합

def csi_age_cat(x):
    if x < 40:
        return '40세미만'
    elif x >= 40 and x < 50:
        return '40-50세'
    elif x >= 50 and x < 60:
        return '50-60세'
    elif x >= 60 and x < 70:
        return '60-70세'
    elif x >= 70:
        return '70세이상'
    
def csi_income_cat(x):
    if x < 1000000:
        return '100만원미만'
    elif x >= 1000000 and x < 2000000:
        return '100-200만원'
    elif x >= 2000000 and x < 3000000:
        return '200-300만원'
    elif x >= 3000000 and x < 4000000:
        return '300-400만원'
    elif x >= 4000000 and x < 5000000:
        return '400-500만원'
    elif x >= 5000000:
        return '500만원이상'

# 병합을 위한 열 생성

tp_master['연령대'] = tp_master['나이'].apply(csi_age_cat)
tp_master['소득수준'] = np.round(tp_master['yearly_income'] / 12).apply(csi_income_cat)

## csi index와 tp_master frame 병합
for csi_col in csi_index.CSI코드.unique():
    
    ### 현재생활형편CSI
    # 병합 임시 DF_성별
    tp_생활_성별 = csi_index[(csi_index.CSI코드 == f'{csi_col}') & ((csi_index.구분기준 == '남자') | (csi_index.구분기준 == '여자'))]

    tp_생활_성별 = tp_생활_성별.rename(columns = {'score': f'{csi_col}_성별',
                              '구분기준': 'gender'}).drop('CSI코드', axis = 1)

    tp_생활_성별.gender = tp_생활_성별.gender.apply(lambda x : 1 if x == '남자' else 0)
    tp_생활_성별.Date = tp_생활_성별.Date.apply(lambda x: x[:7])

    
    ## 나이별
    tp_생활_나이별 = csi_index[(csi_index.CSI코드 == f'{csi_col}') & ((csi_index.구분기준 == '40세미만') | (csi_index.구분기준 == '40-50세') \
                                                  | (csi_index.구분기준 == '50-60세') | (csi_index.구분기준 == '60-70세') | (csi_index.구분기준 == '70세이상'))]

    tp_생활_나이별 = tp_생활_나이별.rename(columns = {'score': f'{csi_col}_나이',
                              '구분기준': '연령대'}).drop('CSI코드', axis = 1)
    tp_생활_나이별.Date = tp_생활_나이별.Date.apply(lambda x: x[:7])

    
    ## 소득별 
    tp_생활_소득별 = csi_index[(csi_index.CSI코드 == f'{csi_col}') & ((csi_index.구분기준 == '100만원미만') | (csi_index.구분기준 == '100-200만원') | (csi_index.구분기준 == '200-300만원')\
                                                   | (csi_index.구분기준 == '300-400만원') | (csi_index.구분기준 == '400-500만원')| (csi_index.구분기준 == '500만원이상'))]

    tp_생활_소득별 = tp_생활_소득별.rename(columns = {'score': f'{csi_col}_소득',
                                            '구분기준': '소득수준'}).drop('CSI코드', axis = 1)
    tp_생활_소득별.Date = tp_생활_소득별.Date.apply(lambda x: x[:7])


    tp_master = pd.merge(tp_master, tp_생활_성별, left_on =['month','gender'], right_on = ['Date', 'gender'], how = 'left').drop('Date',axis = 1)
    tp_master = pd.merge(tp_master, tp_생활_나이별, left_on =['month','연령대'], right_on = ['Date', '연령대'], how = 'left').drop('Date',axis = 1)
    tp_master = pd.merge(tp_master, tp_생활_소득별, left_on =['month','소득수준'], right_on = ['Date', '소득수준'], how = 'left').drop('Date',axis = 1)

    
    # 해당 유저의 항목별 CSI
    tp_master[f'{csi_col}'] = np.round((tp_master[f'{csi_col}_성별'] +  tp_master[f'{csi_col}_나이'] + tp_master[f'{csi_col}_소득']) / 3)

# 소비심리지수 생성
tp_master['소비심리지수'] = np.round((tp_master['현재생활형편CSI'] + tp_master['현재경기판단CSI'] +  tp_master['생활형편전망CSI'] +\
tp_master['소비지출전망CSI'] + tp_master['가계수입전망CSI'] + tp_master['향후경기전망CSI']) / 6, 2)

## 해당 코드를 돌리면 필요 없는 열은 삭제됨

tp_master.drop(columns = ['나이', '연령계층별','현재생활형편CSI', '현재경기판단CSI_성별', '현재경기판단CSI_나이', '현재경기판단CSI_소득',
       '현재경기판단CSI', '생활형편전망CSI_성별', '생활형편전망CSI_나이', '생활형편전망CSI_소득',
       '생활형편전망CSI', '향후경기전망CSI_성별', '향후경기전망CSI_나이', '향후경기전망CSI_소득',
       '향후경기전망CSI', '가계수입전망CSI_성별', '가계수입전망CSI_나이', '가계수입전망CSI_소득',
       '가계수입전망CSI', '소비지출전망CSI_성별', '소비지출전망CSI_나이', '소비지출전망CSI_소득',
       '소비지출전망CSI', '현재가계부채CSI_성별', '현재가계부채CSI_나이', '현재가계부채CSI_소득',
     '가계부채전망CSI_성별', '가계부채전망CSI_나이', '가계부채전망CSI_소득','price_index' ]
      , axis = 1, inplace=True )

target = tp_master.is_applied
tp_master.drop(['is_applied'],axis='columns',inplace=True)
tp_master['is_applied'] = target

tp_master.isna().sum()

tp_master.info()

## Deployment csv

tp_master.to_csv('../Data/Master_df.csv',index=False)
