==> 5.5. Count Value Generation.ipynb <==
# <font color='CC3D3D'> 0.5_5 Count_Value_Generation
    
- log_data.csv로부터 개별 유저가 하루 동안 발생시킨 총 이벤트의 개수와 대출 신청과 관련된 이벤트의 개수를 집계
- 이벤트 총 개수 <span style="color:blue"> **event_cnt.csv** </span> 생성
- 대출 신청 관련 이벤트 총 개수 <span style="color:blue"> **loan_apply_cnt.csv** </span> 생성
- 둘을 합친 데이터 <span style="color:blue"> **cnt_merge.csv** </span> 생성

# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import pickle
from IPython.display import display
import gc
import warnings
import time

warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)
PATH = "../input/bcgdataset"

def view_info(df:pd.DataFrame) -> None:
    print(df.shape)
    print()
    df.info()
    print()
    display(df.head())

# Data Load

loan = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/loan_result.csv", engine='c',
                   low_memory=True,
                   parse_dates=['loanapply_insert_time'],
                   dtype={'application_id': 'uint32', 'bank_id': 'uint8', 'product_id': 'uint8'},
                   usecols=['application_id', 'loanapply_insert_time', 'is_applied'])
log = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/log_data.csv", engine='c', 
                  dtype={'user_id': 'uint32', 'event': 'object'},
                  usecols=['user_id', 'event', 'timestamp', 'date_cd'], 
                  low_memory=True,
                  parse_dates=['timestamp', 'date_cd'])
user = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/user_spec.csv", engine='c',
                   dtype={'application_id': 'uint32', 'user_id': 'uint32'},
                   low_memory=True,
                   parse_dates=['insert_time'],
                   usecols=['application_id', 'user_id', 'insert_time'])

view_info(loan)
print()
view_info(log)
print()
view_info(user)

# Data Structure

app_user = {user['application_id'].iloc[i]: user['user_id'].iloc[i] for i in range(user.shape[0])}
loan['user_id'] = loan['application_id'].map(app_user)
loan.dropna(subset=['user_id'], inplace=True)
loan['user_id'] = loan['user_id'].astype(np.uint32)
loan['date_cd'] = pd.to_datetime(loan['loanapply_insert_time'].dt.date)
loan.sort_values(by=['user_id', 'loanapply_insert_time'], inplace=True)
loan.reset_index(drop=True, inplace=True)
view_info(loan)

del user
gc.collect()

## apply 2번 이상한 사람들
total_apply = loan[['user_id', 'application_id', 'date_cd']].drop_duplicates()
total_apply = total_apply.groupby(by=['user_id', 'date_cd']).count()

multi_apply = total_apply[total_apply['application_id'] > 1]
multi_apply.sort_values(by=['user_id', 'date_cd'], inplace=True)
multi_apply.reset_index(drop=False, inplace=True)
view_info(multi_apply)

single_apply = total_apply[total_apply['application_id'] == 1]
single_apply.sort_values(by=['user_id', 'date_cd'], inplace=True)
single_apply.reset_index(drop=False, inplace=True)
view_info(single_apply)

multi_loan = loan[(loan['user_id'].isin(multi_apply['user_id'].unique())) & (loan['date_cd'].isin(multi_apply['date_cd'].unique()))]
multi_loan['hour'] = multi_loan['loanapply_insert_time'].dt.hour

multi_time = multi_loan[['user_id', 'date_cd', 'hour']].drop_duplicates()
display(multi_time)
print()

mt_group = multi_time[['user_id', 'date_cd', 'hour']].groupby(['user_id', 'date_cd']).count()
mt_group.reset_index(drop=False, inplace=True)
mt_group[mt_group['hour'] == 1]

only_one = mt_group[mt_group['hour'] == 1]
for i, (user, date_cd)in enumerate(zip(only_one['user_id'], only_one['date_cd'])):
    display(loan[(loan['user_id'] == user) & (loan['date_cd'] == date_cd)])
    print()
    if i == 4:
        break

log = log[log['user_id'].isin(loan['user_id'].unique())]
log.sort_values(by=['user_id', 'timestamp'], inplace=True)
log.reset_index(drop=True, inplace=True)
log.info()
print()
log.head()

loan[loan['user_id'] == 65].head(50)

log[log['user_id'] == 65]

# Generate

* 로그 데이터가 무지 엉망이다 -> loan apply 시간이랑 맞지도 않고, application별로 구분되지도 않는다
* 따라서 방법은 다음과 같다.
    - 일단 유저의 일간 duration을 구하고, loan 데이터에 붙인다.
    - 결측치 처리 방법은 다음 중 하나
        - train과 test를 다르게 처리한다면, train에 대해서는 apply한 날과 그렇지 않은 날의 user별 duration의 평균을 각각 구해 그 평균으로 채워주고, test에 대해서는 '기대출 수가 많은 사람일수록 신청 확률이 높다'는 가정에 근거해 대출 신청한 application 비율에 따른 가중 평균으로 채워준다.
        - train과 test를 같은 방법으로 처리한다면, 위에 적은 방법 중 후자를 택한다.

del date_cd, i, mt_group, multi_apply, multi_loan, multi_time, only_one, single_apply, total_apply, user
gc.collect()

## Events

log['month'] = log['date_cd'].dt.month

app_event = log[['user_id', 'event', 'date_cd']].groupby(by=['user_id', 'date_cd']).count()
app_event.reset_index(drop=False, inplace=True)
display(app_event.head())

### Deployment CSV

app_event.to_csv('../Data/event_cnt.csv', index=False)

time.sleep(5)

user_subdfs = []
for user in tqdm(log['user_id'].unique()):
    subdf = log[log['user_id'] == user]
    subdf.sort_values(by=['user_id', 'timestamp'], inplace=True)
    user_subdfs.append(subdf)

threshold = len(user_subdfs) // 5
listA = user_subdfs[:threshold]
listB = user_subdfs[threshold:2*threshold]
listC = user_subdfs[2*threshold:3*threshold]
listD = user_subdfs[3*threshold:4*threshold]
listE = user_subdfs[4*threshold:]

def loanpply_cnt(array:list) -> pd.DataFrame:
    result_df = pd.DataFrame(columns=['user_id', 'date_cd', 'loan_cnt'])
    visit_dt = subdf['date_cd'].unique()
    for item in tqdm(array):
        user_id = item['user_id'].unique()[0]
        for date_cd in item['date_cd'].unique():
            subsubdf = item[item['date_cd'] == date_cd]
            loan_cnt = subsubdf[subsubdf['event'].str.contains('LoanApply')].shape[0]
            result_df = result_df.append({'user_id': np.uint32(user_id), 
                                          'date_cd': pd.to_datetime(date_cd), 
                                          'loan_cnt': np.uint16(loan_cnt)}, ignore_index=True)
    return result_df

resultA = loanpply_cnt(listA)
resultB = loanpply_cnt(listB)
resultC = loanpply_cnt(listC)
resultD = loanpply_cnt(listD)
resultE = loanpply_cnt(listE)

loan_apply_count = pd.concat([resultA, resultB, resultC, resultD, resultE])

# ## loanapply관련 이벤트 발생 횟수
# loan_apply_count = pd.DataFrame(columns=['user_id', 'date_cd', 'loan_cnt'])
# for user_id in tqdm(log['user_id'].unique()):
#     subdf = log[log['user_id'] == user_id]
#     for date_cd in subdf['date_cd'].unique():
#         subsubdf = subdf[subdf['date_cd'] == date_cd]
#         loan_cnt = subsubdf[subsubdf['event'].str.contains('LoanApply')].shape[0]
#         loan_apply_count = loan_apply_count.append({'user_id': np.uint32(user_id),
#                                                     'date_cd': pd.to_datetime(date_cd),
#                                                     'loan_cnt': np.uint8(loan_cnt)}, ignore_index=True)

### Deployment CSV

loan_apply_count.to_csv('../Data/loan_apply_cnt.csv', index=False)

cnt_merge = pd.merge(left=app_event, right=loan_apply_count, on=['user_id', 'date_cd'], how='left')
cnt_merge.fillna(0)
view_info(cnt_merge)

## Deployment CSV

cnt_merge.to_csv('../Data/cnt_merge.csv', index=False)
