==> 5.6. Log Extract App Path.ipynb <==
# <font color='CC3D3D'> 0.5_6 Log_Extract_App_Path
    
- log_data.csv, loan_result.csv, user_spec.csv로부터 개별 유저가 하루 동안 앱에 머물면서 발생시킨 이벤트의 순서를 경로로 정의
- 개별 유저가 방문한 경로 데이터를 집계한 뒤 리스트의 형태로 붙인 데이터
- 대출 신청을 실제로 한 유저의 경로 데이터 <span style="color:blue"> **app_path.csv** </span> 생성
- 대출 신청을 실제로 하지 않은 유저의 경로 데이터 <span style="color:blue"> **.noapp_path.csv** </span> 생성

# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import pickle
from IPython.display import display
import gc

pd.set_option('display.max_columns', None)

# with open(f"{PATH}/빅콘테스트/dtypes.pkl", 'rb') as f:
#     dtypes = pickle.load(f)

def view_info(df:pd.DataFrame) -> None:
    print(df.shape)
    print()
    df.info()
    print()
    display(df.head())

# Data Load

loan = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/loan_result.csv", engine='c',
                   low_memory=True,
                   parse_dates=['loanapply_insert_time'],
                   dtype={'application_id': 'uint32', 'bank_id': 'uint8', 'product_id': 'uint8'},
                   usecols=['application_id', 'loanapply_insert_time', 'is_applied'])
log = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/log_data.csv", engine='c', 
                  dtype={'user_id': 'uint32', 'event': 'object'},
                  usecols=['user_id', 'event', 'timestamp', 'date_cd'], 
                  low_memory=True,
                  parse_dates=['timestamp', 'date_cd'])
user = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/user_spec.csv", engine='c',
                   dtype={'application_id': 'uint32', 'user_id': 'uint32'},
                   low_memory=True,
                   parse_dates=['insert_time'],
                   usecols=['application_id', 'user_id', 'insert_time'])

view_info(loan)
print()
view_info(log)
print()
view_info(user)

loan.isnull().sum()

# Data Structure

app_user = {user['application_id'].iloc[i]: user['user_id'].iloc[i] for i in range(user.shape[0])}
# train = loan[loan['loanapply_insert_time'].dt.month < 6]
# train['user_id'] = train['application_id'].map(app_user)
# train.dropna(subset=['user_id'], inplace=True)
# train['user_id'] = train['user_id'].astype(np.uint32)

# view_info(train)

train = loan[loan['loanapply_insert_time'].dt.month < 6]
apply_group = train[['application_id', 'is_applied']].groupby(by='application_id').sum()
apply_group.reset_index(drop=False, inplace=True)
apply_group.columns = ['application_id', 'total_applied']

view_info(apply_group)

app_id = apply_group[apply_group['total_applied'] != 0]['application_id'].unique().astype(np.uint32)
noapp_id = apply_group[apply_group['total_applied'] == 0]['application_id'].unique().astype(np.uint32)

print(f"# of application ids that applied: {app_id.shape[0]}")
print(f"# of application ids that not applied: {noapp_id.shape[0]}")

print(app_id.max())
print(noapp_id.max())

total_loan = pd.merge(left=train, right=apply_group, on='application_id', how='inner')
view_info(total_loan)

total_loan['user_id'] = total_loan['application_id'].map(app_user)
total_loan.dropna(subset=['user_id'], inplace=True)
total_loan['user_id'] = total_loan['user_id'].astype(np.uint32)
total_loan['date_cd'] = pd.to_datetime(total_loan['loanapply_insert_time'].dt.date)
view_info(total_loan)

log_train = log[log['date_cd'].dt.month < 6]
view_info(log_train)

app_log_merge = pd.merge(left=log_train, right=total_loan[['user_id', 'application_id', 'date_cd', 'total_applied']], on=['user_id', 'date_cd'], how='inner')
app_log_merge['total_applied'] = app_log_merge['total_applied'].astype(np.uint8)
view_info(app_log_merge)

actual_log = app_log_merge[app_log_merge['application_id'].isin(app_id)]
no_log = app_log_merge[app_log_merge['application_id'].isin(noapp_id)]

view_info(actual_log)

view_info(no_log)

del app_log_merge
gc.collect()

def get_path(df:pd.DataFrame) -> pd.DataFrame:
    df.sort_values(by=['application_id', 'timestamp'], inplace=True)
    event_tuples = dict()
    for appid in tqdm(df['application_id'].unique()):
        subdf = df[df['application_id'] == appid]
        for dt in subdf['date_cd'].unique():
            subsubdf = subdf[subdf['date_cd'] == dt]
            events = tuple(subsubdf['event'])
            if events not in event_tuples:
                event_tuples[events] = 1
            else:
                event_tuples[events] += 1

    result = pd.DataFrame({'path': event_tuples.keys(), 'count': event_tuples.values()})

    return result

## Deployment CSV

app_path = get_path(actual_log)
app_path.to_csv("../Data/path_applied_user.csv", index=False)

noapp_path = get_path(no_log)
noapp_path.to_csv("../Data/path_no_applied_user.csv", index=False)

app_path.head(10)

app_path.sort_values(by='count', ascending=False, inplace=True)
app_path.reset_index(drop=True, inplace=True)
for i in range(20):
    path, count = app_path.head(20).iloc[i]
    print(f"{path}\t{count}")

noapp_path.head(10)

noapp_path.sort_values(by='count', ascending=False, inplace=True)
noapp_path.reset_index(drop=True, inplace=True)
for i in range(20):
    path, count = noapp_path.head(20).iloc[i]
    print(f"{path}\t{count}")
