==> 6.1. Data Split Cluster.ipynb <==
# <font color='CC3D3D'> Data Split Cluster
    
### 1. Data Merge
- <span style="color:green">**Master_df_v2_cluster.csv** </span>와 <span style="color:green">**log_feature_clansing.csv** </span>를 Merge
    
### 2. Data Split
- Merge한 데이터를 Train Test로 나눔
- Train 데이터를 Train, Validation 데이터로 나눔
- Validation 데이터를 Public, Private로 데이터를 나눔

### 3. Deployment CSV
- 나눠진 데이터를 <span style="color:blue">**X_train_cluster.csv, X_public_cluster.csv, X_private_cluster.csv, X_test_cluster.csv** </span>를 생성

# Import

import pandas as pd
from sklearn.model_selection import train_test_split

# Data Load

master_df = pd.read_csv('../Data/Master_df_v2_cluster.csv')

log_feature = pd.read_csv('../Data/log_feature_clansing.csv')

master_df.isna().sum()

master_df.info()

log_feature.isna().sum()

log_feature.info()

# Data Mergem

master_df = pd.concat([master_df,log_feature],axis=1)

master_df.isna().sum()

master_df.info()

# Missing Value

master_df.segment.value_counts()

master_df.event.fillna(master_df.event.mean(),inplace=True)
master_df.loan_cnt.fillna(master_df.loan_cnt.mean(),inplace=True)
master_df.duration.fillna(master_df.duration.mean(),inplace=True)
master_df.visit.fillna(master_df.visit.mean(),inplace=True)
master_df.segment.fillna(0,inplace=True)

master_df.isna().sum()

master_target = master_df['is_applied']
master_df.drop(['is_applied'],axis=1,inplace=True)
master_df['is_applied'] = master_target

master_df.info()

# Data Split

## Train Test

train_data = master_df[master_df.is_applied.notna()].reset_index(drop=True)
test_data = master_df[master_df.is_applied.isna()].reset_index(drop=True)

train_data.isna().sum()

test_data.isna().sum()

columns = train_data.columns.tolist()

train_data

test_data

# Train Validation Split

data = train_data.drop(['is_applied'],axis=1)
target = train_data['is_applied']

target

data

x_train, x_valid, y_train, y_valid = train_test_split(data, target, test_size=0.3, shuffle=True, stratify=target, random_state=34)

print(x_train.shape)
print(y_train.shape)

print(x_valid.shape)
print(y_valid.shape)

y_train.isna().sum()

y_valid.isna().sum()

# Public Private Split

x_public, x_private, y_public, y_private = train_test_split(x_valid, y_valid, test_size=0.7, shuffle=True, stratify=y_valid, random_state=34)

print(x_public.shape)
print(y_public.shape)

print(x_private.shape)
print(y_private.shape)

y_public.isna().sum()

y_private.isna().sum()

# Data Structure

## Train

x_train['is_applied'] = y_train
x_train = x_train.reset_index(drop=True)
x_train

## Public

x_public['is_applied'] = y_public
x_public = x_public.reset_index(drop=True)
x_public

## Private

x_private['is_applied'] = y_private
x_private = x_private.reset_index(drop=True)
x_private

## Test

test_data

# Info

x_train.isna().sum()

x_public.isna().sum()

x_private.isna().sum()

test_data.isna().sum()

# Deployment CSV

x_train.to_csv('../Data/X_train_cluster.csv',index=False)
x_public.to_csv('../Data/X_public_cluster.csv',index=False)
x_private.to_csv('../Data/X_private_cluster.csv',index=False)
test_data.to_csv('../Data/X_test_cluster.csv',index=False)
