==> 6.4. Undersampling+SMOTE.ipynb <==
# <font color='CC3D3D'> 0.6_4 Undersampling+SMOTE
    
- 딥러닝 수행하기 전 불균형성을 맞추기 위해 데이터의 40%를 undersampling한 뒤 SMOTE 알고리즘을 사용해 희소 label의 양을 늘려 label 균형을 맞춰주는 데이터를 생성하는 코드
- 불균형 해소 데이터 <span style="color:blue"> **oversample_train.csv** </span> 생성

# Import

!pip install -q imblearn
print("Done!")

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
from tqdm import tqdm
import os
import pickle
import warnings
import gc

from sklearn.model_selection import train_test_split
from imblearn.over_sampling import RandomOverSampler, SMOTE

RANDOM_STATE = 4158
pd.set_option('display.max_columns', None)
CHUNKSIZE = 50000

with open(f"{PATH}/dtypes.pkl", 'rb') as f:
    dtypes = pickle.load(f)

warnings.filterwarnings('ignore')

def loaddata(fname:str, chunksize:int, dtype:dict=None, columns:list=None):  
    df = pd.DataFrame()
    for chunk in tqdm(pd.read_csv(fname, engine='python', low_memory=True, chunksize=chunksize, dtype=dtype)):
        df = pd.concat([df, chunk], axis=0)
        del chunk
        gc.collect()

    return df

# Data Load

train = loaddata("../Data/master_train_data.csv", chunksize=CHUNKSIZE, dtype=dtypes)
valid = loaddata("../Data/master_public_data.csv", chunksize=CHUNKSIZE, dtype=dtypes)

# Undersampling & SMOTE

_, under = train_test_split(train, test_size=0.4, shuffle=True, random_state=RANDOM_STATE, stratify=train['is_applied'])

under.shape

under['is_applied'].value_counts() / under['is_applied'].value_counts().sum()

train['is_applied'].value_counts() / train['is_applied'].value_counts().sum()

X_under = under.drop('is_applied', axis=1)
y_under = under['is_applied']

smote = SMOTE(random_state=RANDOM_STATE, n_jobs=-1)
X_over, y_over = smote.fit_resample(X_under, y_under)

X_over.shape

y_over.value_counts()

master_train_oversample = pd.concat([X_over, y_over], axis=1)

master_train_oversample.head()

## Deployment

master_train_oversample.to_csv("../Data/oversample_train.csv", index=False)
