==> 6.5. UnderSampling Train Test Split.ipynb <==
# <font color='CC3D3D'> 0.6_5 UnderSampling Train Test Split
    
- **master_train_data.csv** 중 30%만 랜덤하게 뽑아서 95:5 비율로 undersampling (0 : 1 = 2,039,032 : 117,670)
- 그 중 80%를 랜덤하게 뽑아서 Train으로 설정 (0 : 1 = 1,631,225 : 94,136)
    - <span style="color:blue"> **master_train_undersample.csv** </span> 생성
- 그 중 20%를 랜덤하게 뽑아서 Test으로 설정 (0 : 1 = 407,807 : 23,534)
    - <span style="color:blue"> **master_test_undersample.csv** </span> 생성

# Import

import numpy as np

import pandas as pd

from sklearn.model_selection import train_test_split

# Data Load

master_train = pd.read_csv('../Data/master_train_data.csv')

master_train

# Undersampling

train = master_train.drop('is_applied', axis=1)
target = master_train.is_applied

X_train, X_valid, y_train, y_valid = train_test_split(train, target, test_size=0.7, 
                                                      random_state=42, shuffle=True, stratify=target)

y_train.value_counts()

y_valid.value_counts()

X_train, X_valid, y_train, y_valid = train_test_split(X_train, y_train, test_size=0.2, 
                                                      random_state=42, shuffle=True, stratify=y_train)

y_train.value_counts()

1631225/(1631225 + 94136)

y_valid.value_counts()

train_undersample = pd.concat([X_train, y_train], axis=1)
test_undersample = pd.concat([X_valid, y_valid], axis=1)

## Deployment

train_undersample.to_csv('../Data/master_train_undersample.csv', index=False)
test_undersample.to_csv('../Data/master_test_undersample.csv', index=False)
