==> 6.6. Oversampling SMOTE.ipynb <==
# <font color='CC3D3D'> 0.6_6 OverSamplinig SMOTE
    
- train 데이터에서 0과 1 비율을 5:5로 맞춰줌 (0 : 1 = 1,631,225 : 1,631,225)
- SMOTE 알고리즘을 사용해 희소 label의 양을 늘려 label 균형을 맞춰주는 데이터를 생성하는 코드
- 불균형 해소 데이터 <span style="color:blue"> **master_train_oversample.csv** </span> 생성

# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.datasets import make_classification
from sklearn.model_selection import cross_validate
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.ensemble import RandomForestClassifier

from imblearn.pipeline import Pipeline
from imblearn.combine import SMOTETomek
from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import TomekLinks

from sklearn.metrics import classification_report, confusion_matrix

import scipy
from sklearn.svm import SVC

# Data Load

train_undersample = pd.read_csv('../Data/master_train_undersample.csv')
test_undersample = pd.read_csv('../Data/master_test_undersample.csv')

X_train = train_undersample.drop('is_applied', axis=1)
y_train = train_undersample['is_applied']

# SMOTE 객체 생성
smote = SMOTE(random_state=42)

# 샘플링
X_train_resampled, y_train_resampled = smote.fit_resample(X_train, y_train)

X_train_resampled

y_train_resampled.value_counts()

master_train_oversample = pd.concat([X_train_resampled, y_train_resampled], axis=1)

## Deployment

master_train_oversample.to_csv('../Data/master_train_oversample.csv', index=False)

pd.read_csv('../Data/master_train_oversample.csv')
