==> 7.1. Pycaret Anomaly.ipynb <==
함유진

0.7_1

- 0.7_1
    
    ```python
    # <font color='CC3D3D'> Pycaret Anomaly
        
    ### 1. 무작위 표본 추출
        빠른 시간 내에 모델의 성능을 평가하기 위해서,
        train, public dataset에서 100,000개 씩 데이터를 추출함.
        
        
    ### 2. 모델 실험
        K-Nearest Neighbor / isolation forest / Minimum Covariance Determinant
        
    - K-Nearest Neighbor
        
        이상치 데이터는 멀리 존재할 것이라는 가정 하에 K개의 근접 이웃까지 거리를 계산함.
        
        이웃과의 거리가 다른 데이터들에 비해 멀다면 이상치일 확률이 높음. 
        
        - <span style="color:green"> **pycaret_knn.pickle** </span> 생성
        - <span style="color:blue"> **0.04310428143423214** </span> public f1_score
        
    - isolation forest
        
        Tree 기반의 이상 탐지 비지도 알고리즘
        
        - <span style="color:green"> **pycaret_iforest.pickle.csv** </span> 생성
        - <span style="color:blue"> *0.1201625386996904** </span> public f1_score
        
    - Minimum Covariance Determinant
        
        MCD는 이상치의 영향을 최소화하는 방향으로 평균과 공분산 행렬을 추정하는 방식임.
    
        - <span style="color:green"> **pycaret_mcd.pickle** </span> 생성
        - <span style="color:blue"> **0.08543499511241447** </span> public f1_score
    ```
    

user_spec cluster 생성

- user_spec cluster 생성
    
    ```python
    # <font color='CC3D3D'> Clustering User Spec 
        
    ### 1. Kmeans
        user_spec의 numeric featuresf를 사용해 kmeans 방법을 사용하여 고객 군집화를 진행함.
        Elbow point, Silhouette score를 복합적으로 사용해 군집의 개수를 5개로 정함.
        
    ### 2. 시각화
        umap.plot을 통해 군집이 잘 생성되었는지 확인함.
    ```
    

user_cluster_analysis

- user_cluster_analysis
    
    ```python
    # <font color='CC3D3D'> User Spec Clusters Analysis
        
    ### 1. Cluster별 특징 파악
        - 신용점수
        - 연소득
        - 희망대출금액
        - 연령
        - 기대출액
        
    ### 2. Cluster별 Event 분석
        - LoanApply 중 이탈률 분석
        - GetCreditInfo 사용 횟수 분석
    ```

# Import

from pycaret.anomaly import *

import pandas as pd

from sklearn.metrics import accuracy_score, precision_score, recall_score, confusion_matrix,f1_score

# Data Load

train = pd.read_csv('../Data/master_train_data.csv') 
public = pd.read_csv('../Data/master_public_data.csv')
private = pd.read_csv('../Data/master_private_data.csv')
test = pd.read_csv('../Data/master_test_data.csv')

train = train.sample(n=100000)
public = public.sample(n=100000)
private = private.sample(n=100000)

train_target = train['is_applied']
public_target = public['is_applied']
# private_target = private['is_applied']

train = train.drop(['is_applied'],axis=1)
public = public.drop(['is_applied'],axis=1)
# private = private.drop(['is_applied'],axis=1)

train.info()

# Pycaret

## Set Up

anomaly = setup(train,
                use_gpu = True,
                session_id = 42,
               )

## KNN

knn = create_model('knn')

print(knn)

public_target = pd.DataFrame(public_target)
public_target['is_appliled'] = public_target['is_applied'].apply(lambda x: '0' if x == 0 else '1')
public_target['is_appliled']

unseen_predictions = predict_model(knn, data=public)
f1_score(public_target['is_applied'],unseen_predictions['Anomaly'])

save_model(knn,'../Model/pycaret_knn')

## Iforest

iforest = create_model('iforest')

unseen_predictions = predict_model(iforest, data=public)
f1_score(public_target['is_applied'],unseen_predictions['Anomaly'])

save_model(iforest,'../Model/pycaret_iforest')

## MCD

mcd = create_model('mcd')

unseen_predictions = predict_model(mcd, data=public)
f1_score(public_target['is_applied'],unseen_predictions['Anomaly'])

save_model(mcd,'../Model/pycaret_mcd')
