==> 7.3. Sampling Pycaret ETRF.ipynb <==
# <font color='CC3D3D'> 0.7_3 Sampling Pycaret ETRF
    
- <span style="color:green"> **master_train_oversample.csv**</span>와 <span style="color:green">**master_public_data.csv** </span> 불러옴
- Oversampling까지 진행한 데이터로 Pycaret 모델 학습
- Public Data F1_Score 기준 0.3 이상인 RandomForest, ExtraTrees 2가지 모델 선정 및 저장

# Import

!pip install pycaret[full]

import pandas as pd
import numpy as np
from pycaret.classification import *

# Data Load

master_train_oversample = pd.read_csv('../Data/master_train_oversample.csv') ; master_train_oversample.head()
master_public_data = pd.read_csv('../Data/master_public_data.csv') ; master_public_data.head()

master_public_label = master_public_data['is_applied']
master_public_data = master_public_data.drop('is_applied', axis=1)

# https://pycaret.readthedocs.io/en/stable/api/classification.html

#cat_feats = ['income_type', 'employment_type', 'houseown_type', 'purpose', 'personal_rehabilitation_yn', 'personal_rehabilitation_complete_yn', 
#             'gender', 'bank_diversity', 'product_diversity', 'holiday', 'weekday', 'foreign', 'segment', 'bank_label', 'product_label'] categorical_features=cat_feats

exp_clf = setup(data = master_train_oversample, target = 'is_applied', train_size = 0.8, data_split_shuffle=True,
                session_id=123, use_gpu = True, test_data = master_public_data)

best_model = compare_models(sort="F1", fold=3)

best_model

save_model(best_model, '../Model/pycaret_et')

prediction = predict_model(best_model, data = master_public_data)

rf = create_model('rf', fold=3)
save_model(rf, '../Model/pycaret_rf')

prediction = predict_model(rf, data = master_public_data)
