==> 8.1. TabNet Predict Sampling.ipynb <==
# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
from tqdm import tqdm
import os
import random
import pickle
import warnings
import gc
import joblib

from sklearn.preprocessing import MinMaxScaler, StandardScaler, LabelEncoder
from sklearn.model_selection import train_test_split, cross_val_score, StratifiedKFold
from sklearn.metrics import (precision_recall_curve, roc_curve, make_scorer, log_loss,
                             f1_score, confusion_matrix, auc, roc_auc_score, accuracy_score)

import torch
from torch import nn
from torch.utils.data import DataLoader, Dataset
import torch.optim as optim
import torch.nn.functional as F
from pytorch_tabnet.tab_model import TabNetClassifier
from pytorch_tabnet.metrics import Metric
from pytorch_tabnet.pretraining import TabNetPretrainer
from torch.optim.lr_scheduler import ReduceLROnPlateau

import optuna
from optuna.samplers import TPESampler

DEVICE = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
RANDOM_STATE = 4158
CHUNKSIZE = 50000

with open(f"./dtypes.pkl", 'rb') as f:
    dtypes = pickle.load(f)

pd.set_option('display.max_columns', None)
warnings.filterwarnings('ignore')
print(f"Using PyTorch version: {torch.__version__}, Device: {DEVICE}")
%matplotlib inline

def loaddata(fname:str, chunksize:int, dtype:dict=None, columns:list=None):  
    df = pd.DataFrame()
    for chunk in tqdm(pd.read_csv(fname, engine='python', low_memory=True, chunksize=chunksize, dtype=dtype)):
        df = pd.concat([df, chunk], axis=0)
        del chunk
        gc.collect()

    return df

# Data Load

private = loaddata("../Data/master_private_data.csv", chunksize=CHUNKSIZE, dtype=dtypes)
test = loaddata("../Data/master_test_data.csv", chunksize=CHUNKSIZE)

public = loaddata("../Data/master_public_data.csv", chunksize=CHUNKSIZE, dtype=dtypes)

cat_feats = ['income_type', 'employment_type', 'houseown_type', 'purpose', 'personal_rehabilitation_yn', 'personal_rehabilitation_complete_yn', 
             'gender', 'bank_diversity', 'product_diversity', 'holiday', 'weekday', 'foreign', 'segment', 'bank_label', 'product_label']
cat_idxs = [i for i, col in enumerate(private.columns.values) if col in cat_feats]

for col in private.columns.values:
    if private[col].dtype == 'bool':
        private[col] = private[col].astype('uint8')
    if test[col].dtype == 'bool':
        test[col] = test[col].astype('uint8')
    if public[col].dtype == 'bool':
        public[col] = public[col].astype('uint8')

for col in cat_feats:
    if private[col].min() != 0:
        private[col] -= private[col].min()
    if test[col].min() != 0:
        test[col] -= test[col].min()
    if public[col].min() != 0:
        public[col] -= public[col].min()

# cat_dims = [[col].nunique() for col in cat_feats]

for i in range(test.shape[0]):
    if test['product_diversity'].iloc[i] == 79:
        test['product_diversity'].iloc[i] = 9

print(f"Unique categories")
for col in cat_feats:
    print(f"\tprivate {col} -> {np.sort(private[col].unique())}", end='')
    print(f"\ttest {col} -> {np.sort(test[col].unique())}")

X_public = public.drop('is_applied', axis=1)
y_public = public['is_applied']

## Evaluation Metric

class F1Score(Metric):
    def __init__(self):
        self._name = "f1"
        self._maximize = True
    
    def __call__(self, y_true, y_score):
        y_score = (y_score[:, 1] > 0.5).astype(np.uint8)
        return f1_score(y_true, y_score)

## Non-sampled

clf = joblib.load(f"../Model/clf_no_sampling.pkl")

public_pred_proba = clf.predict_proba(X_public.values)

private_pred_proba = clf.predict_proba(private.values)

test_pred_proba = clf.predict_proba(test.values)

public_pred = (public_pred_proba[:, 1] >= 0.5).astype(np.uint8)
private_pred = (private_pred_proba[:, 1] >= 0.5).astype(np.uint8)
test_pred = (test_pred_proba[:, 1] >= 0.5).astype(np.uint8)

## Deployment CSV

public_pred = pd.DataFrame({'predict': public_pred})
private_pred = pd.DataFrame({'predict': private_pred})
test_pred = pd.DataFrame({'predict': test_pred})

public_pred.to_csv('../Data/public_predict.csv', index=False)
private_pred.to_csv('../Data/private_predict.csv', index=False)
test_pred.to_csv('../Data/test_predict.csv', index=False)

## Sampled

clf2 = joblib.load(f"../Model/clf_smote_pretrain.pkl")

public_pred_proba2 = clf2.predict_proba(X_public.values)

private_pred_proba2 = clf2.predict_proba(private.values)

test_pred_proba2 = clf2.predict_proba(test.values)

public_pred2 = (public_pred_proba2[:, 1] >= 0.5).astype(np.uint8)
private_pred2 = (private_pred_proba2[:, 1] >= 0.5).astype(np.uint8)
test_pred2 = (test_pred_proba2[:, 1] >= 0.5).astype(np.uint8)

public_pred2 = pd.DataFrame({'predict': public_pred2})
private_pred2 = pd.DataFrame({'predict': private_pred2})
test_pred2 = pd.DataFrame({'predict': test_pred2})

public_pred2.to_csv('../Data/tabnet_public_predict_sam.csv', index=False)
private_pred2.to_csv('../Data/tabnet_private_predict_sam.csv', index=False)
test_pred2.to_csv('../Data/tabnet_test_predict_sam.csv', index=False)
