==> 9. Ensemble.ipynb <==
# Import

import pandas as pd
import numpy as np
import joblib
import pickle

from sklearn.ensemble import IsolationForest
from pycaret import anomaly as py_an
from pycaret import classification as py_cl
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier

import torch
from torch import nn
from torch.utils.data import DataLoader, Dataset
import torch.optim as optim
import torch.nn.functional as F

from pytorch_tabnet.tab_model import TabNetClassifier
from pytorch_tabnet.metrics import Metric
from pytorch_tabnet.pretraining import TabNetPretrainer
from torch.optim.lr_scheduler import ReduceLROnPlateau

from sklearn.metrics import f1_score

DEVICE = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
RANDOM_STATE = 4158

# Def

class F1Score(Metric):
    def __init__(self):
        self._name = "f1"
        self._maximize = True
    
    def __call__(self, y_true, y_score):
        y_score = (y_score[:, 1] >= 0.5).astype(np.uint8)
        return f1_score(y_true, y_score)

def get_pred_label(model_pred):
    # IsolationForest 모델 출력 (1:정상, -1:불량(사기)) 이므로 (0:정상, 1:불량(사기))로 Label 변환
    model_pred = np.where(model_pred == 1, 0, model_pred)
    model_pred = np.where(model_pred == -1, 1, model_pred)
    return model_pred

# Data Load

public = pd.read_csv('../Data/master_public_data.csv',)
private = pd.read_csv('../Data/master_private_data.csv')
private_y = pd.read_csv('../Data/master_correct_answer_data.csv')['is_applied'].values
test = pd.read_csv('../Data/master_test_data.csv')

print(public.shape)
print(private.shape)
print(private_y.shape)
print(test.shape)

public_x = public.drop(['is_applied'],axis=1)
public_y = public['is_applied'].values

test_x = test.drop(['is_applied'],axis=1)
test_y = test['is_applied'].values

# Model Load

## Anomaly

an_iforest_base = joblib.load('../Model/Isolationforest_base_1012.pkl')

an_isolation_pycaret = py_an.load_model('../Model/pycaret_iforest')

an_knn_pycaret = joblib.load('../Model/pycaret_knn.pkl')

an_mcd_pycaret = py_an.load_model('../Model/pycaret_mcd')

## ML

ml_lgbm = joblib.load('../Model/LGBM_Optuna_model_0.44_WithLog.pkl')

ml_cat = joblib.load('../Model/Cat_Optuna_model_0.48_withLog.pkl')

ml_et = py_cl.load_model('../Model/pycaret_et')

ml_rf = py_cl.load_model('../Model/pycaret_rf')

# Anomaly

an_iforest_base_public_pred = an_iforest_base.predict(public_x)
an_iforest_base_priate_pred = an_iforest_base.predict(private)
an_iforest_base_test_pred = an_iforest_base.predict(test_x)

an_isolation_pycaret_public_pred = py_an.predict_model(an_isolation_pycaret, data = public_x)
an_isolation_pycaret_private_pred = py_an.predict_model(an_isolation_pycaret, data = private)
an_isolation_pycaret_test_pred = py_an.predict_model(an_isolation_pycaret, data = test_x)

an_knn_pycaret_public_pred = py_an.predict_model(an_knn_pycaret, data = public_x)
an_knn_pycaret_private_pred = py_an.predict_model(an_knn_pycaret, data = private)
an_knn_pycaret_test_pred = py_an.predict_model(an_knn_pycaret, data = test_x)

an_mcd_pycaret_public_pred = py_an.predict_model(an_mcd_pycaret, data = public_x)
an_mcd_pycaret_private_pred = py_an.predict_model(an_mcd_pycaret, data = private)
an_mcd_pycaret_test_pred = py_an.predict_model(an_mcd_pycaret, data = test_x)

## ML

ml_lgbm_public_pred = ml_lgbm.predict(public_x)
ml_lgbm_private_pred = ml_lgbm.predict(private)
ml_lgbm_test_pred = ml_lgbm.predict(test_x)

ml_cat_public_pred = ml_cat.predict(public_x)
ml_cat_private_pred = ml_cat.predict(private)
ml_cat_test_pred = ml_cat.predict(test_x)

ml_et_public_pred = py_cl.predict_model(ml_et, data = public_x)
ml_et_private_pred = py_cl.predict_model(ml_et, data = private)
ml_et_test_pred = py_cl.predict_model(ml_et, data = test_x)

ml_rf_public_pred = py_cl.predict_model(ml_rf, data = public_x)
ml_rf_private_pred = py_cl.predict_model(ml_rf, data = private)
ml_rf_test_pred = py_cl.predict_model(ml_rf, data = test_x)

## Deployment CSV

### Anomaly

pd.DataFrame(an_iforest_base_public_pred).to_csv('../Data/an_isolation_public_pred.csv',index=False)
pd.DataFrame(an_iforest_base_priate_pred).to_csv('../Data/an_isolation_private.csv',index=False)
pd.DataFrame(an_iforest_base_test_pred).to_csv('../Data/an_isolation_test.csv',index=False)

pd.DataFrame(an_isolation_pycaret_public_pred).to_csv('../Data/an_isolation_pycaret_public_pred.csv',index=False)
pd.DataFrame(an_isolation_pycaret_private_pred).to_csv('../Data/an_isolation_pycaret_private_pred.csv',index=False)
pd.DataFrame(an_isolation_pycaret_test_pred).to_csv('../Data/an_isolation_pycaret_test_pred.csv',index=False)

pd.DataFrame(an_knn_pycaret_public_pred).to_csv('../Data/an_knn_pycaret_public_pred.csv',index=False)
pd.DataFrame(an_knn_pycaret_private_pred).to_csv('../Data/an_knn_pycaret_private_pred.csv',index=False)
pd.DataFrame(an_knn_pycaret_test_pred).to_csv('../Data/an_knn_pycaret_test_pred.csv',index=False)

an_mcd_pycaret_public_pred.to_csv('../Data/an_mcd_ptcaret_public_pred.csv',index=False)
an_mcd_pycaret_private_pred.to_csv('../Data/an_mcd_ptcaret_private_pred.csv',index=False)
an_mcd_pycaret_test_pred.to_csv('../Data/an_mcd_ptcaret_tset_pred.csv',index=False)

### ML

pd.DataFrame(ml_lgbm_public_pred).to_csv('../Data/ml_lgbm_public_pred.csv',index=False)
pd.DataFrame(ml_lgbm_private_pred).to_csv('../Data/ml_lgbm_private_pred.csv',index=False)
pd.DataFrame(ml_lgbm_test_pred).to_csv('../Data/ml_lgbm_test_pred.csv',index=False)

pd.DataFrame(ml_cat_public_pred).to_csv('../Data/ml_cat_public_pred_csv.',index=False)
pd.DataFrame(ml_cat_private_pred).to_csv('../Data/ml_cat_private_pred_csv.',index=False)
pd.DataFrame(ml_cat_test_pred).to_csv('../Data/ml_cat_test_pred_csv.',index=False)

ml_et_public_pred.to_csv('../Data/ml_et_public_pred.csv',index=False)
ml_et_priate_pred.to_csv('../Data/ml_et_private_pred.csv',index=False)
ml_et_test_pred.to_csv('../Data/ml_et_test_pred.csv',index=False)

ml_rf_public_pred.to_csv('../Data/ml_rf_public_pred.csv',index=False)
ml_rf_private_pred.to_csv('../Data/ml_rf_private_pred.csv',index=False)
ml_rf_test_pred.to_csv('../Data/ml_rf_test_pred.csv',index=False)

## Data Load

### Data

public_y = pd.read_csv('../Data/master_public_data.csv',usecols=['is_applied'])
private_y = pd.read_csv('../Data/master_correct_answer_data.csv')['is_applied'].values
test = pd.read_csv('../Data/master_test_data.csv')

### Anomaly

an_iforest_base_public_pred = pd.read_csv('../Data/an_isolation_public_pred.csv')
an_iforest_base_private_pred = pd.read_csv('../Data/an_isolation_private.csv')
an_iforest_base_test_pred = pd.read_csv('../Data/an_isolation_test.csv')

an_isolation_pycaret_public_pred = pd.read_csv('../Data/an_isolation_pycaret_public_pred.csv',usecols=['Anomaly'])
an_isolation_pycaret_private_pred = pd.read_csv('../Data/an_isolation_pycaret_private_pred.csv',usecols=['Anomaly'])
an_isolation_pycaret_test_pred = pd.read_csv('../Data/an_isolation_pycaret_test_pred.csv',usecols=['Anomaly'])

an_knn_pycaret_public_pred = pd.read_csv('../Data/an_knn_pycaret_public_pred.csv',usecols=['Anomaly'])
an_knn_pycaret_private_pred = pd.read_csv('../Data/an_knn_pycaret_private_pred.csv',usecols=['Anomaly'])
an_knn_pycaret_test_pred = pd.read_csv('../Data/an_knn_pycaret_test_pred.csv',usecols=['Anomaly'])

an_mcd_pycaret_public_pred = pd.read_csv('../Data/an_mcd_ptcaret_public_pred.csv',usecols=['Anomaly'])
an_mcd_pycaret_private_pred = pd.read_csv('../Data/an_mcd_ptcaret_private_pred.csv',usecols=['Anomaly'])
an_mcd_pycaret_test_pred = pd.read_csv('../Data/an_mcd_ptcaret_tset_pred.csv',usecols=['Anomaly'])

### ML

ml_lgbm_public_pred = pd.read_csv('../Data/ml_lgbm_public_pred.csv')
ml_lgbm_private_pred = pd.read_csv('../Data/ml_lgbm_private_pred.csv')
ml_lgbm_test_pred = pd.read_csv('../Data/ml_lgbm_test_pred.csv')

ml_cat_public_pred = pd.read_csv('../Data/ml_cat_public_pred_csv.')
ml_cat_private_pred = pd.read_csv('../Data/ml_cat_private_pred_csv.')
ml_cat_test_pred = pd.read_csv('../Data/ml_cat_test_pred_csv.')

ml_et_public_pred = pd.read_csv('../Data/ml_et_public_pred.csv',usecols=['Label'])
ml_et_priate_pred = pd.read_csv('../Data/ml_et_private_pred.csv',usecols=['Label'])
ml_et_test_pred = pd.read_csv('../Data/ml_et_test_pred.csv',usecols=['Label'])

ml_rf_public_pred = pd.read_csv('../Data/ml_rf_public_pred.csv',usecols=['Label'])
ml_rf_private_pred = pd.read_csv('../Data/ml_rf_private_pred.csv',usecols=['Label'])
ml_rf_test_pred = pd.read_csv('../Data/ml_rf_test_pred.csv',usecols=['Label'])

### Deep

dp_tabnet_public_pred = pd.read_csv('../Data/tabnet_public_predict.csv')
dp_tabnet_private_pred = pd.read_csv('../Data/tabnet_private_predict.csv')
dp_tabnet_test_pred = pd.read_csv('../Data/tabnet_test_predict.csv')

dp_sam_tabnet_public_pred = pd.read_csv('../Data/tabnet_public_predict_sam.csv')
dp_sam_tabnet_private_pred = pd.read_csv('../Data/tabnet_private_predict_sam.csv')
dp_sam_tabnet_test_pred = pd.read_csv('../Data/tabnet_test_predict_sam.csv')

# F1

## Anomaly

an_iforest_base_public_pred = get_pred_label(an_iforest_base_public_pred)
an_iforest_base_private_pred = get_pred_label(an_iforest_base_private_pred)

print(f1_score(an_iforest_base_public_pred,public_y))
print(f1_score(an_iforest_base_private_pred,private_y))

print(f1_score(an_isolation_pycaret_public_pred,public_y))
print(f1_score(an_isolation_pycaret_private_pred,private_y))

print(f1_score(an_knn_pycaret_public_pred,public_y))
print(f1_score(an_knn_pycaret_private_pred,private_y))

print(f1_score(an_mcd_pycaret_public_pred,public_y))
print(f1_score(an_mcd_pycaret_private_pred,private_y))

## ML

print(f1_score(ml_lgbm_public_pred,public_y))
print(f1_score(ml_lgbm_private_pred,private_y))

print(f1_score(ml_cat_public_pred,public_y))
print(f1_score(ml_cat_private_pred,private_y))

print(f1_score(pd.to_numeric(ml_et_public_pred['Label']),public_y))
print(f1_score(pd.to_numeric(ml_et_priate_pred['Label']),private_y))

print(f1_score(pd.to_numeric(ml_rf_public_pred['Label']),public_y))
print(f1_score(pd.to_numeric(ml_rf_private_pred['Label']),private_y))

## Deep

print(f1_score(dp_sam_tabnet_public_pred,public_y))
print(f1_score(dp_sam_tabnet_private_pred,private_y))

print(f1_score(dp_tabnet_public_pred,public_y))
print(f1_score(dp_tabnet_private_pred,private_y))

# Ensemble

## Public

### Anomaly

an_public_pred = an_iforest_base_public_pred | an_isolation_pycaret_public_pred | an_knn_pycaret_public_pred | an_mcd_pycaret_public_pred

print(f1_score(an_public_pred,public_y))

an_public_pred = an_iforest_base_public_pred | an_isolation_pycaret_public_pred  | an_mcd_pycaret_public_pred

print(f1_score(an_public_pred,public_y))

### ML

ml_tree_public_pred = pd.to_numeric(ml_et_public_pred['Label']).astype(np.uint8) | pd.to_numeric(ml_rf_public_pred['Label']).astype(np.uint8)

print(f1_score(ml_tree_public_pred,public_y))

ml_boost_public_pred = pd.DataFrame(ml_lgbm_public_pred).astype(np.uint8) | pd.DataFrame(ml_cat_public_pred).astype(np.uint8)

print(f1_score(ml_boost_public_pred,public_y))

ml_public_pred = np.zeros(shape=ml_tree_public_pred.shape)
for i in range(len(ml_public_pred)):
    ml_public_pred[i] = (ml_tree_public_pred[i] | ml_boost_public_pred.values[i]).astype(np.uint8)

print(f1_score(ml_public_pred,public_y))

### Deep

dp_public_pred = dp_sam_tabnet_public_pred | dp_tabnet_public_pred

print(f1_score(dp_public_pred,public_y))

### ALL

an_isolation_pycaret_public_pred.values.flatten()

ml_cat_public_pred.values.flatten()

all_pred = an_isolation_pycaret_public_pred.values.flatten().astype(int) | ml_cat_public_pred.values.flatten().astype(int) | dp_sam_tabnet_public_pred.values.flatten().astype(int)

print(f1_score(all_pred,public_y))

pred_ens = []
for a,b,c,d in zip(ml_lgbm_public_pred.values.flatten().astype(int),
                      ml_cat_public_pred.values.flatten().astype(int), 
                      dp_tabnet_public_pred.values.flatten().astype(int),
                      ml_et_public_pred['Label'].values.flatten().astype(int)):
    if (a+b+c+d) >= 2:
        pred_ens.append(1)
    else:
        pred_ens.append(0)

np.array(pred_ens)

print(f1_score(np.array(pred_ens),public_y))

## Private

pred_ens = []
for a,b,c in zip(ml_lgbm_private_pred.values.flatten().astype(int),
                      ml_cat_private_pred.values.flatten().astype(int), 
                      dp_tabnet_private_pred.values.flatten().astype(int)):
    if (a+b+c) >= 2:
        pred_ens.append(1)
    else:
        pred_ens.append(0)

np.array(pred_ens)

print(f1_score(np.array(pred_ens),private_y))

## Test

pred_ens = []
for a,b,c in zip(ml_lgbm_test_pred.values.flatten().astype(int),
                      ml_cat_test_pred.values.flatten().astype(int), 
                      dp_tabnet_test_pred.values.flatten().astype(int)):
    if (a+b+c) >= 2:
        pred_ens.append(1)
    else:
        pred_ens.append(0)

np.array(pred_ens)

len(pred_ens)

sum(pred_ens)

## Deployment

submission = pd.read_csv('../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/데이터분석분야_퓨처스부문_평가데이터.csv')

submission

submission['is_applied'] = pred_ens
submission

submission.to_csv('../../평가데이터/데이터분석분야_퓨처스부문_롱Loan팀_평가데이터.csv',index=False)
