==> 11.1. Log Path Pivoting And Ngrams.ipynb <==
# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import pickle
from IPython.display import display
import gc
import warnings
import nltk
from nltk.util import ngrams, everygrams
from nltk.probability import ConditionalProbDist, MLEProbDist

nltk.download('punkt')
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

with open("../Data/dtypes.pkl", 'rb') as f:
    dtypes = pickle.load(f)

def viewinfo(df:pd.DataFrame) -> None:
    print(df.shape)
    print()
    df.info()
    print()
    display(df.head())

# Pivot Table Generation

log = pd.read_csv("../Data/2022빅콘테스트_데이터분석리그_데이터분석분야_퓨처스부문_데이터셋_220908/log_data.csv", engine='c', low_memory=True, parse_dates=['timestamp', 'date_cd'])
viewinfo(log)

* Main events: ViewLoanApplyIntro & StartLoanApply / EndLoanApply / UseLoanManage / UsePrepayCalc / UseDSRCalc / GetCreditInfo

log.sort_values(by=['user_id', 'timestamp'], inplace=True)
log.head(10)

main_events = ['ViewLoanApplyIntro', 'StartLoanApply', 'UserLoanManage', 'UsePrepayCalc', 'UseDSRCalc', 'GetCreditInfo']
# log.drop(['mp_os', 'mp_app_version'], axis=1, inplace=True)

def user_separate(df:pd.DataFrame, key:str) -> list:
    app_subdfs = []
    for app in tqdm(df[key].unique()):
        subdf = df[df[key] == app]
        app_subdfs.append(subdf)
        
    return app_subdfs

def get_event(array:list) -> pd.DataFrame:
    event_log = pd.DataFrame(columns=['user_id', 'events'])
    for item in tqdm(array):
        user_id = item['user_id'].unique().item()
        main_ = []
        for date_cd in item['date_cd'].unique():
            subdf = item[item['date_cd'] == date_cd]
            if subdf[subdf['event'].isin(main_events)].empty:
                main_.append('ETC')
            else:
                main_ += list(subdf[subdf['event'].isin(main_events)]['event'].unique())
        
        event_log = event_log.append({'user_id': np.uint32(user_id), 'events': main_}, ignore_index=True)
    
    return event_log

user_arrays = user_separate(log, key='user_id')

threshold = len(user_arrays) // 10
listA = user_arrays[:threshold]
listB = user_arrays[threshold:2*threshold]
listC = user_arrays[2*threshold:3*threshold]
listD = user_arrays[3*threshold:4*threshold]
listE = user_arrays[4*threshold:5*threshold]
listF = user_arrays[5*threshold:6*threshold]
listG = user_arrays[6*threshold:7*threshold]
listH = user_arrays[7*threshold:8*threshold]
listI = user_arrays[8*threshold:9*threshold]
listJ = user_arrays[9*threshold:]

resultA = get_event(listA)
viewinfo(resultA)

resultB = get_event(listB)
resultC = get_event(listC)
resultD = get_event(listD)
resultE = get_event(listE)
resultF = get_event(listF)
resultG = get_event(listG)
resultH = get_event(listH)
resultI = get_event(listI)
resultJ = get_event(listJ)

result = pd.concat([resultA, resultB, resultC, resultD, resultE, resultF, resultG, resultH, resultI, resultJ], axis=0)
result.reset_index(drop=True, inplace=True)

viewinfo(result)

result.to_csv("../Data/main_events.csv", index=False)

# N-grams

* references
    - [N-gram 모델 구현하기 기초 | Python, NLTK](https://seanpark11.tistory.com/89?category=962465)
    - [확률론적 언어 모형](https://datascienceschool.net/03%20machine%20learning/03.01.05%20%ED%99%95%EB%A5%A0%EB%A1%A0%EC%A0%81%20%EC%96%B8%EC%96%B4%20%EB%AA%A8%ED%98%95.html)

event_dict = {
    'SignUp': 'sign up',
    'OpenApp': 'open app',
    'Login': 'login',
    'ViewLoanApplyIntro': 'view loan apply intro',
    'StartLoanApply': 'start loan apply',
    'CompleteIDCertification': 'complete id certification',
    'EndLoanApply': 'end loan apply',
    'UseLoanManage': 'use loan manage',
    'UsePrepayCalc': 'use prepay calculator',
    'UseDSRCalc': 'use debt service ratio calculator',
    'GetCreditInfo': 'get credit information'
}

main_events = ['ViewLoanApplyIntro', 'StartLoanApply', 'UserLoanManage', 'UsePrepayCalc', 'UseDSRCalc', 'GetCreditInfo']

df = pd.read_csv("../Data/main_events.csv")
df.head()

type(df['events'].iloc[0])

df['events'] = df['events'].map(lambda x: x[1:-1])
df['events'] = df['events'].map(lambda x: [item[1:-1].replace("'", "") for item in x.split(', ')])

df.head()

type(df['events'].iloc[0])

df['num_events'] = df['events'].map(lambda x: len(x))
df.head()

df['num_events'].value_counts()

## People with more than 3 tags

df['event_sent'] = df['events'].map(lambda x: ' '.join(x))

to_ngram = df[df['num_events'] >= 3]
no_ngram = df[df['num_events'] < 3]

print(no_ngram['num_events'].max())
print(no_ngram['num_events'].min())

to_ngram.reset_index(drop=True, inplace=True)

to_ngram['trigrams'] = to_ngram['events'].map(lambda x: ngrams(x, n=3))
to_ngram.head(10)

to_ngram['trigrams_list'] = to_ngram['events'].map(lambda x: list(ngrams(x, n=3)))
to_ngram.head(10)

to_ngram['dist'] = to_ngram['trigrams'].map(lambda x: nltk.ConditionalFreqDist([((t[0], t[1]), t[2]) for t in x]))
to_ngram

to_ngram['prob'] = to_ngram['dist'].map(lambda x: ConditionalProbDist(x, MLEProbDist))
to_ngram

i = 0
dist = to_ngram['dist'].iloc[i]
last_prev, last = to_ngram['events'].iloc[0][-2:]

cpd = ConditionalProbDist(dist, MLEProbDist)
print(cpd)
print(cpd[(last_prev, last)])

for item in main_events:
    print(f"Prob of {item} given {last_prev, last} -> {cpd[(last_prev, last)].prob(item)}")

maximal_action = []
minimal_action = []
for i in tqdm(range(to_ngram.shape[0])):
    dist = to_ngram['dist'].iloc[i]
    last_prev, last = to_ngram['events'].iloc[i][-2:]

    cpd = to_ngram['prob'].iloc[i]
    scores = np.array([cpd[(last_prev, last)].prob(item) for item in main_events])

    max_idx = np.where(scores == scores.max())[0]
    max_item = np.random.choice(np.array(main_events)[max_idx])

    try:
        min_idx = np.where((scores == scores.min()) & (scores > 0))[0]
        min_item = np.random.choice(np.array(main_events)[min_idx])
    except:
        min_idx = np.where((scores == scores.min()))[0]
        min_item = np.random.choice(np.array(main_events)[min_idx])        

    maximal_action.append(max_item)
    minimal_action.append(min_item)
    
to_ngram['maximal_action'] = maximal_action
to_ngram['minimal_action'] = minimal_action

to_ngram

to_ngram['maximal_action'].value_counts()

to_ngram['minimal_action'].value_counts()

to_ngram.to_csv("../Data/ngram_.csv", index=False)

Datano_ngram.to_csv("../Data/no_ngram_.csv", index=False)
